Move MigrationJob to migration package.
